requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
certifi==2022.9.24
cffi==1.15.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
distro==1.9.0
dj-database-url==2.3.0
Django==4.2.18
django-allauth==65.3.1
django-autocomplete-light==3.11.0
django-autoslug==1.9.9
django-cities-light==3.10.1
django-crispy-forms==2.3
django-widget-tweaks==1.5.0
exceptiongroup==1.2.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.4
jiter==0.8.2
jwt==1.3.1
msgpack==1.1.0
oauthlib==3.2.2
openai==1.64.0
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.5
Pillow==9.2.0
progressbar2==4.5.0
psycopg==3.2.4
psycopg2==2.9.10
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.6.0
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dotenv==0.19.0
python-utils==3.9.1
python3-openid==3.2.0
pytz==2022.5
redis==5.2.1
requests==2.28.1
requests-oauthlib==1.3.1
routing==0.2.0
sniffio==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tqdm==4.67.1
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==1.26.12
daphne>=3.0.0
uvicorn
uvicorn[standard]
websockets
django-crontab-0.7.1


